+++ web-common/src/components/menu/DashboardMetricsDraggableList.svelte
